Two changes to speed up fe55 cluster fitting: vectorize call to pixel_integral and seed fit with cluster moments.

On a small test case I used these two changes:
1) give very similar results,
2) reduce the total time spend in the fitter from 71 seconds to 21 seconds,
3) reduce the amount of time spent evaluating the PSF integral from 51 seconds to 14 seconds.